Refactor Makefile and generate proto files
Refactor Makefile to include a new target called "proto" that generates proto files from overlay package. The proto files are generated from request.go and response.go files in the overlay package. The generated proto files are then moved to the overlay directory and the package name is updated to "mpesaoverlay.overlay". Additionally, the data type "uint8" is replaced with "uint32" in the generated proto files. Finally, the protoc command is used to compile the proto files into Go code.

The following changes were made in this commit:

- Added a new target "proto" in the Makefile to generate proto files
- Generated proto files from request.go and response.go in the overlay package
- Moved the generated proto files to the overlay directory
- Updated the package name in the generated proto files to "mpesaoverlay.overlay"
- Replaced the data type "uint8" with "uint32" in the generated proto files
- Compiled the proto files into Go code using the protoc command.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>