Add metrics package and create metrics for request count and latency.
This commit adds a new package `metrics` to the codebase. It includes a new file `metrics.go` which defines functions for creating metrics related to request count and latency. The `MakeMetrics` function initializes and returns instances of `Counter` and `Summary` from the `go-kit/metrics/prometheus` package. The `Counter` represents the number of requests received, while the `Summary` represents the latency of the requests.

The `MakeMetrics` function takes two parameters: `namespace` and `subsystem`, which are used to set the namespace and subsystem for the metrics. The `Counter` is created using the `NewCounterFrom` function, which takes `CounterOpts` as input. The `CounterOpts` specify the name and help message for the counter. The `Summary` is created using the `NewSummaryFrom` function, which also takes `SummaryOpts` as input.

The `Counter` and `Summary` are returned as pointers from the `MakeMetrics` function.

This commit is necessary to add metrics tracking to the codebase and enable monitoring of request count and latency.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>