Refactor Dockerfile and docker-compose.yml
- Refactor Dockerfile and Dockerfile.dev to include the source URL in the LABEL section
- Update docker-compose.yml to use the latest version of the overlay image

The changes made in this commit refactor the Dockerfile and docker-compose.yml files. The Dockerfile and Dockerfile.dev now include the source URL in the LABEL section. Additionally, the docker-compose.yml file has been updated to use the latest version of the overlay image.

These changes improve the maintainability and traceability of the project by providing a clear reference to the source code repository in the Docker images.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>