Add Docker workflow for building and pushing Docker images
This commit adds a new Docker workflow file, `docker.yaml`, which sets up a GitHub Actions workflow for building and pushing Docker images. The workflow is triggered on pushes to the `main` branch.

The workflow consists of several steps:
1. Checking out the code using the `actions/checkout` action.
2. Setting up QEMU using the `docker/setup-qemu-action` action.
3. Setting up Docker Buildx using the `docker/setup-buildx-action` action.
4. Logging in to the GitHub Container Registry using the `docker/login-action` action.
5. Building and pushing the Docker image using the `docker/build-push-action` action.

The Dockerfile used for building the image is located in the `./docker` directory. The image is built for both the `linux/amd64` and `linux/arm64` platforms.

These changes enable automated building and pushing of Docker images for the project.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>