Fix Dockerfile build process and add go mod tidy and go mod vendor commands
The Dockerfile was updated to include the commands "go mod tidy" and "go mod vendor" to ensure that the project's Go modules are properly managed. This will help to ensure that the project's dependencies are up to date and consistent. Additionally, the build process was fixed to correctly build the specified service and move the resulting binary to the appropriate location.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>