feat(ci): Add build job to CI workflow
This commit adds a new build job to the CI workflow. The build job is responsible for building the project using different versions of Go (1.20.x and 1.21.x) and on different operating systems (Ubuntu, macOS, and Windows). This will ensure that the project is tested and built successfully on multiple platforms.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>