feat(ci): Add golangci-lint step to CI workflow
This commit adds a new step to the CI workflow in the `.github/workflows/ci.yaml` file. The new step, `golangci-lint`, is responsible for running the golangci-lint tool to perform static code analysis and linting on the codebase. This step helps ensure code quality and adherence to best practices. The step is executed using the `curl` command to download the necessary script from the internet and then run it.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>