feat(ci): update golangci-lint installation command
The previous installation command for golangci-lint in the ci.yaml file was not correctly setting the installation path. This commit updates the command to ensure that golangci-lint is installed in the proper location by specifying the installation path using the "-b" flag. This change ensures that the linting step in the CI workflow runs successfully.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>