feat(docker): rmove go mod tidy step to Dockerfile
This commit removes the `go mod tidy` step to the Dockerfile in the `docker` directory. Previously, this step was included, which caused issues with the dependency management. Now, when building the Docker image, the `go mod tidy` command will be executed, ensuring that all dependencies are properly managed. This will help to avoid any potential issues related to missing or outdated dependencies.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>