feat(mqtt): add Hook struct and related methods
This commit adds a new file, hook.go, which contains the implementation of the Hook struct. The Hook struct is used to handle MQTT hooks and includes methods for serving MQTT requests, logging, and handling JSON payloads. The Hook struct extends the mqtt.HookBase struct and includes a reference to the mqtt.Server and a logger from the go.uber.org/zap package.

The Hook struct is intended to be used in conjunction with the mochi-mqtt server library (github.com/mochi-mqtt/server/v2) and the mpesaoverlay package (github.com/0x6flab/mpesaoverlay).

This commit is a new feature addition to the codebase and is expected to enhance the functionality of the MQTT module.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>