feat(ci): add docker build step to CI workflow
This commit adds a new step to the CI workflow in the `.github/workflows/ci.yaml` file. The new step runs the `make dockers` command, which builds the Docker images for the application. This addition ensures that the Docker images are built and available as part of the CI process, allowing for easier deployment and testing.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>