feat(cmd/grpc): add fatal log message when failing to create service
Previously, when creating the service in the main.go file, an error message was logged using the "Error" method. This commit changes that to use the "Fatal" method instead. This ensures that if there is an error creating the service, the program will exit immediately with a fatal log message.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>