feat(ci): seperate linting and ci
This commit adds a new linting job to the CI workflow. The linting job runs on the latest Ubuntu environment and installs Go. It then installs the golangci-lint tool and runs the linting process on the codebase. This ensures that the code follows best practices and catches any potential issues early on.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>