feat: Add .goreleaser.yaml configuration file
Add a new file .goreleaser.yaml with some sensible defaults for the project. The file includes hooks for go mod tidy and go generate ./..., as well as a build configuration for the main CLI command.

This commit introduces the necessary configuration file for Goreleaser, which will be used for future releases and deployments.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>