feat(build): add exclusion rules to changelog filters
This commit adds two exclusion rules to the changelog filters in the `.goreleaser.yaml` file. The first rule excludes any commits with the prefix "chore(deps):", and the second rule excludes any commits starting with "Bump". These rules are added to ensure that certain types of commits are not included in the generated changelog.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>