feat(.goreleaser.yaml): Add changelog configuration
Add the changelog configuration to the .goreleaser.yaml file. This configuration includes sorting the changelog in ascending order and applying filters to exclude certain commit messages from being included in the changelog. Additionally, it defines a group for dependency updates, which will be used to categorize relevant commits in the changelog.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>