feat: Add CI and Docker workflow badges to README
This commit adds CI and Docker workflow badges to the README file. The CI badge indicates the status of the CI workflow, while the Docker badge indicates the status of the Docker workflow. These badges provide visibility into the current state of the workflows and help track the progress of the project.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>