feat(release): Add release workflow
Add a new release workflow to automate the process of creating releases. This workflow is triggered when a tag starting with "v" is pushed. It runs on the latest version of Ubuntu and sets the necessary environment variables for Docker. This workflow will handle the release process and permissions for writing contents, id-token, and packages.

The release workflow file has been added at `.github/workflows/release.yml`.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>