feat: Add new environment variables for HTTP and MQTT ports
This commit adds new environment variables for the HTTP and MQTT ports in the example.env file. The HTTP port is set to 80 and the HTTPS port is set to 443. Additionally, the MQTT port is set to 1883 and the MQTTS port is set to 8883.

These changes allow for easier configuration of the ports used by the core services in the application.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>