feat(Makefile): add volume flag to docker-compose down command
The previous version of the Makefile did not include the volume flag when running the `docker-compose down` command. This caused the associated volumes to persist even after stopping the containers.

This commit adds the `-v` flag to the `docker-compose down` command in the Makefile, ensuring that the volumes associated with the containers are also removed when stopping the application.

This change improves the development workflow by providing a clean slate when restarting the application.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>