feat(grafana): Add Loki datasource and example dashboard
Added a new YAML file `datasources.yaml` in the `docker/grafana` directory to configure the Loki datasource for Grafana. The datasource is named "Loki" and is of type "loki". The URL is set to "http://loki:3100" and it is set as the default datasource.

Also added a new JSON file `example.json` in the same directory to provide an example dashboard configuration.

These changes enable the use of Loki as a datasource in Grafana and provide an example dashboard for reference.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>