feat(grafana): add dashboard.yaml configuration file
This commit adds a new file, `dashboard.yaml`, to the `docker/grafana` directory. The `dashboard.yaml` file contains the configuration for the Grafana dashboard provider. It includes the necessary API version, provider details, and options such as the path for the dashboards.

The `dashboard.yaml` file is used to configure the Grafana dashboard provider to read dashboards from the specified path (`/var/lib/grafana/dashboards`). The provider is set to update every 10 seconds and does not allow UI updates. Deletion of dashboards is enabled.

These changes enable the Grafana dashboard provider to load dashboards from the specified path and keep them up to date with the latest changes.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>