feat(docs): update deployment guide
This commit updates the deployment guide in the docs folder. Specifically, it removes the mention of Traefik as a reverse proxy and updates the information about essential components like Loki and Prometheus for log monitoring and observability. The changes ensure that the guide provides accurate and up-to-date instructions for deploying the application to the cloud.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>