feat(cli): add b2b command
This commit adds a new file `b2b.go` to the `cli` package. The `b2b.go` file contains the implementation of the `b2b` command, which allows users to perform Business-to-Business transactions using the MpesaOverlay library. The `b2b` command utilizes the `mpesa` package to interact with the Mpesa API and perform the necessary operations for B2B transactions.

The `b2b.go` file is licensed under the Apache-2.0 license and includes the necessary copyright and license information at the top of the file.

This new feature will enhance the functionality of the CLI tool by providing users with the ability to perform B2B transactions easily.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>